refactor(agent-core): split runtime lifecycle helpers out of lib (#1682)
