+++ crates/tau-agent-core/src/lib.rs
@@ -11,18 +11,39 @@ use std::{
 };
 
 use async_trait::async_trait;
-use jsonschema::validator_for;
 use serde_json::{json, Value};
 use tau_ai::{
     ChatRequest, ChatUsage, LlmClient, Message, MessageRole, StreamDeltaHandler, TauAiError,
     ToolCall, ToolChoice, ToolDefinition,
 };
-use tau_memory::runtime::{cosine_similarity, embed_text_vector};
 pub use tau_safety::{
     DefaultLeakDetector, DefaultSanitizer, LeakDetector, SafetyMode, SafetyPolicy, SafetyStage,
     Sanitizer,
 };
 use thiserror::Error;
+
+mod runtime_safety_memory;
+mod runtime_startup;
+mod runtime_tool_bridge;
+mod runtime_turn_loop;
+
+pub(crate) use runtime_safety_memory::{assistant_text_suggests_failure, retrieve_memory_matches};
+pub(crate) use runtime_startup::{
+    cache_insert_with_limit, lock_or_recover, normalize_direct_message_content,
+    sleep_with_cancellation, spawn_async_event_handler_worker,
+};
+pub(crate) use runtime_tool_bridge::execute_tool_call_inner;
+#[cfg(test)]
+pub(crate) use runtime_turn_loop::extract_json_payload;
+pub(crate) use runtime_turn_loop::{
+    bounded_messages, build_structured_output_retry_prompt, collapse_whitespace,
+    estimate_chat_request_tokens, estimate_usage_cost_usd, is_retryable_ai_error,
+    normalize_cost_alert_thresholds, parse_structured_output, role_label,
+    stream_retry_buffer_on_delta, timeout_duration_from_ms, truncate_chars,
+    StreamingRetryBufferState,
+};
+#[cfg(test)]
+pub(crate) use tau_memory::runtime::embed_text_vector;
 
 /// Public struct `AgentConfig` used across Tau components.
 ///
@@ -2012,813 +2033,6 @@ impl Agent {
         });
         result.is_error
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct ChatRequestTokenEstimate {
-    input_tokens: u32,
-    total_tokens: u32,
-}
-
-fn estimate_chat_request_tokens(request: &ChatRequest) -> ChatRequestTokenEstimate {
-    let message_tokens = request.messages.iter().fold(0u32, |acc, message| {
-        acc.saturating_add(estimate_message_tokens(message))
-    });
-    let tool_tokens = request.tools.iter().fold(0u32, |acc, tool| {
-        acc.saturating_add(estimate_tool_definition_tokens(tool))
-    });
-    let input_tokens = message_tokens.saturating_add(tool_tokens).saturating_add(2);
-    let total_tokens = input_tokens.saturating_add(request.max_tokens.unwrap_or(0));
-
-    ChatRequestTokenEstimate {
-        input_tokens,
-        total_tokens,
-    }
-}
-
-fn estimate_message_tokens(message: &Message) -> u32 {
-    let mut total = 4u32;
-    for block in &message.content {
-        match block {
-            tau_ai::ContentBlock::Text { text } => {
-                total = total.saturating_add(estimate_text_tokens(text));
-            }
-            tau_ai::ContentBlock::ToolCall {
-                id,
-                name,
-                arguments,
-            } => {
-                total = total.saturating_add(estimate_text_tokens(id));
-                total = total.saturating_add(estimate_text_tokens(name));
-                total = total.saturating_add(estimate_json_tokens(arguments));
-                total = total.saturating_add(4);
-            }
-            tau_ai::ContentBlock::Image { source } => {
-                total = total.saturating_add(estimate_media_source_tokens(source));
-                total = total.saturating_add(8);
-            }
-            tau_ai::ContentBlock::Audio { source } => {
-                total = total.saturating_add(estimate_media_source_tokens(source));
-                total = total.saturating_add(8);
-            }
-        }
-    }
-    if let Some(tool_call_id) = &message.tool_call_id {
-        total = total.saturating_add(estimate_text_tokens(tool_call_id));
-    }
-    if let Some(tool_name) = &message.tool_name {
-        total = total.saturating_add(estimate_text_tokens(tool_name));
-    }
-    total
-}
-
-fn estimate_tool_definition_tokens(definition: &ToolDefinition) -> u32 {
-    let mut total = 12u32;
-    total = total.saturating_add(estimate_text_tokens(&definition.name));
-    total = total.saturating_add(estimate_text_tokens(&definition.description));
-    total = total.saturating_add(estimate_json_tokens(&definition.parameters));
-    total
-}
-
-fn estimate_json_tokens(value: &Value) -> u32 {
-    let rendered = serde_json::to_string(value).unwrap_or_else(|_| value.to_string());
-    estimate_text_tokens(&rendered)
-}
-
-fn estimate_media_source_tokens(source: &tau_ai::MediaSource) -> u32 {
-    match source {
-        tau_ai::MediaSource::Url { url } => estimate_text_tokens(url),
-        tau_ai::MediaSource::Base64 { mime_type, data } => estimate_text_tokens(mime_type)
-            .saturating_add((data.len() as u32).saturating_div(3))
-            .saturating_add(2),
-    }
-}
-
-fn estimate_text_tokens(text: &str) -> u32 {
-    if text.is_empty() {
-        return 0;
-    }
-    let chars = u32::try_from(text.chars().count()).unwrap_or(u32::MAX);
-    chars.saturating_add(3) / 4
-}
-
-fn estimate_usage_cost_usd(
-    usage: &ChatUsage,
-    input_cost_per_million: Option<f64>,
-    output_cost_per_million: Option<f64>,
-) -> f64 {
-    let input = input_cost_per_million
-        .unwrap_or(0.0)
-        .max(0.0)
-        .mul_add(usage.input_tokens as f64, 0.0)
-        / 1_000_000.0;
-    let output = output_cost_per_million
-        .unwrap_or(0.0)
-        .max(0.0)
-        .mul_add(usage.output_tokens as f64, 0.0)
-        / 1_000_000.0;
-    input + output
-}
-
-fn normalize_cost_alert_thresholds(thresholds: &[u8]) -> Vec<u8> {
-    let mut normalized = thresholds
-        .iter()
-        .copied()
-        .filter(|threshold| (1..=100).contains(threshold))
-        .collect::<Vec<_>>();
-    if normalized.is_empty() {
-        normalized.push(100);
-    }
-    normalized.sort_unstable();
-    normalized.dedup();
-    normalized
-}
-
-fn bounded_messages(messages: &[Message], max_messages: usize) -> Vec<Message> {
-    if max_messages == 0 || messages.len() <= max_messages {
-        return messages.to_vec();
-    }
-
-    if max_messages < 3 {
-        return bounded_messages_without_summary(messages, max_messages);
-    }
-
-    if matches!(
-        messages.first().map(|message| message.role),
-        Some(MessageRole::System)
-    ) {
-        let tail_keep = max_messages - 2;
-        let tail_start = messages.len().saturating_sub(tail_keep);
-        if tail_start <= 1 {
-            return messages.to_vec();
-        }
-
-        let dropped = &messages[1..tail_start];
-        if dropped.is_empty() {
-            return bounded_messages_without_summary(messages, max_messages);
-        }
-
-        let mut bounded = Vec::with_capacity(max_messages);
-        bounded.push(messages[0].clone());
-        bounded.push(Message::system(summarize_dropped_messages(dropped)));
-        bounded.extend_from_slice(&messages[tail_start..]);
-        bounded
-    } else {
-        let tail_keep = max_messages - 1;
-        let tail_start = messages.len().saturating_sub(tail_keep);
-        if tail_start == 0 {
-            return messages.to_vec();
-        }
-
-        let dropped = &messages[..tail_start];
-        if dropped.is_empty() {
-            return bounded_messages_without_summary(messages, max_messages);
-        }
-
-        let mut bounded = Vec::with_capacity(max_messages);
-        bounded.push(Message::system(summarize_dropped_messages(dropped)));
-        bounded.extend_from_slice(&messages[tail_start..]);
-        bounded
-    }
-}
-
-fn bounded_messages_without_summary(messages: &[Message], max_messages: usize) -> Vec<Message> {
-    if max_messages == 0 || messages.len() <= max_messages {
-        return messages.to_vec();
-    }
-
-    if max_messages > 1
-        && matches!(
-            messages.first().map(|message| message.role),
-            Some(MessageRole::System)
-        )
-    {
-        let tail_keep = max_messages - 1;
-        let tail_start = messages.len().saturating_sub(tail_keep);
-        if tail_start <= 1 {
-            return messages.to_vec();
-        }
-        let mut bounded = Vec::with_capacity(max_messages);
-        bounded.push(messages[0].clone());
-        bounded.extend_from_slice(&messages[tail_start..]);
-        bounded
-    } else {
-        messages[messages.len() - max_messages..].to_vec()
-    }
-}
-
-fn summarize_dropped_messages(messages: &[Message]) -> String {
-    let mut user_count = 0usize;
-    let mut assistant_count = 0usize;
-    let mut tool_count = 0usize;
-    let mut system_count = 0usize;
-    let mut excerpts = Vec::new();
-
-    for message in messages {
-        match message.role {
-            MessageRole::User => user_count = user_count.saturating_add(1),
-            MessageRole::Assistant => assistant_count = assistant_count.saturating_add(1),
-            MessageRole::Tool => tool_count = tool_count.saturating_add(1),
-            MessageRole::System => system_count = system_count.saturating_add(1),
-        }
-
-        let content = collapse_whitespace(&message.text_content());
-        if content.is_empty() {
-            continue;
-        }
-        if message.role == MessageRole::System && content.starts_with(CONTEXT_SUMMARY_PREFIX) {
-            continue;
-        }
-        if excerpts.len() >= CONTEXT_SUMMARY_MAX_EXCERPTS {
-            continue;
-        }
-
-        let preview = truncate_chars(&content, CONTEXT_SUMMARY_SNIPPET_MAX_CHARS);
-        excerpts.push(format!("- {}: {}", role_label(message.role), preview));
-    }
-
-    let mut summary = format!(
-        "{CONTEXT_SUMMARY_PREFIX}\n\
-         summarized_messages={}; roles: user={}, assistant={}, tool={}, system={}.",
-        messages.len(),
-        user_count,
-        assistant_count,
-        tool_count,
-        system_count
-    );
-
-    if !excerpts.is_empty() {
-        let excerpt_block = excerpts.join("\n");
-        summary.push_str("\nexcerpts:\n");
-        summary.push_str(&excerpt_block);
-    }
-
-    truncate_chars(&summary, CONTEXT_SUMMARY_MAX_CHARS)
-}
-
-fn role_label(role: MessageRole) -> &'static str {
-    match role {
-        MessageRole::User => "user",
-        MessageRole::Assistant => "assistant",
-        MessageRole::Tool => "tool",
-        MessageRole::System => "system",
-    }
-}
-
-fn collapse_whitespace(text: &str) -> String {
-    text.split_whitespace().collect::<Vec<_>>().join(" ")
-}
-
-fn truncate_chars(text: &str, max_chars: usize) -> String {
-    let total_chars = text.chars().count();
-    if total_chars <= max_chars {
-        return text.to_string();
-    }
-    if max_chars == 0 {
-        return String::new();
-    }
-    if max_chars == 1 {
-        return "…".to_string();
-    }
-
-    let truncate_at = text
-        .char_indices()
-        .nth(max_chars - 1)
-        .map(|(index, _)| index)
-        .unwrap_or(text.len());
-    let mut truncated = text[..truncate_at].to_string();
-    truncated.push('…');
-    truncated
-}
-
-fn assistant_text_suggests_failure(text: &str) -> bool {
-    let normalized = collapse_whitespace(&text.to_lowercase());
-    if normalized.trim().is_empty() {
-        return true;
-    }
-    FAILURE_SIGNAL_PHRASES
-        .iter()
-        .any(|phrase| normalized.contains(phrase))
-}
-
-#[derive(Debug, Clone)]
-struct MemoryEmbeddingApiConfig {
-    model: String,
-    api_base: String,
-    api_key: String,
-}
-
-async fn retrieve_memory_matches(
-    history: &[Message],
-    query: &str,
-    limit: usize,
-    dimensions: usize,
-    min_similarity: f32,
-    config: &AgentConfig,
-) -> Vec<MemoryRecallMatch> {
-    if limit == 0 {
-        return Vec::new();
-    }
-    let candidates = history
-        .iter()
-        .filter_map(|message| match message.role {
-            MessageRole::User | MessageRole::Assistant => {
-                let text = message.text_content();
-                if text.trim().is_empty() {
-                    None
-                } else {
-                    Some((message.role, text))
-                }
-            }
-            MessageRole::Tool | MessageRole::System => None,
-        })
-        .collect::<Vec<_>>();
-    if candidates.is_empty() {
-        return Vec::new();
-    }
-
-    let api_embeddings = if let Some(api_config) = resolve_memory_embedding_api_config(config) {
-        let mut inputs = Vec::with_capacity(candidates.len().saturating_add(1));
-        inputs.push(query.to_string());
-        inputs.extend(candidates.iter().map(|(_, text)| text.clone()));
-        match embed_text_vectors_via_api(&inputs, dimensions, &api_config).await {
-            Ok(vectors) if vectors.len() == inputs.len() => Some(vectors),
-            _ => None,
-        }
-    } else {
-        None
-    };
-
-    let (query_embedding, candidate_embeddings) = if let Some(vectors) = api_embeddings {
-        let query_embedding = vectors.first().cloned().unwrap_or_default();
-        let candidate_embeddings = vectors.into_iter().skip(1).collect::<Vec<_>>();
-        (query_embedding, candidate_embeddings)
-    } else {
-        let query_embedding = embed_text_vector(query, dimensions);
-        let candidate_embeddings = candidates
-            .iter()
-            .map(|(_, text)| embed_text_vector(text, dimensions))
-            .collect::<Vec<_>>();
-        (query_embedding, candidate_embeddings)
-    };
-    if query_embedding.iter().all(|component| *component == 0.0) {
-        return Vec::new();
-    }
-
-    let mut matches = candidates
-        .into_iter()
-        .zip(candidate_embeddings.into_iter())
-        .filter_map(|((role, text), candidate_embedding)| {
-            let score = cosine_similarity(&query_embedding, &candidate_embedding);
-            if score >= min_similarity {
-                Some(MemoryRecallMatch { score, role, text })
-            } else {
-                None
-            }
-        })
-        .collect::<Vec<_>>();
-
-    matches.sort_by(|left, right| right.score.total_cmp(&left.score));
-    matches.truncate(limit);
-    matches
-}
-
-fn resolve_memory_embedding_api_config(config: &AgentConfig) -> Option<MemoryEmbeddingApiConfig> {
-    let model = config.memory_embedding_model.as_deref()?.trim();
-    let api_key = config.memory_embedding_api_key.as_deref()?.trim();
-    if model.is_empty() || api_key.is_empty() {
-        return None;
-    }
-    let api_base = config
-        .memory_embedding_api_base
-        .as_deref()
-        .unwrap_or("https://api.openai.com/v1")
-        .trim()
-        .trim_end_matches('/')
-        .to_string();
-    if api_base.is_empty() {
-        return None;
-    }
-    Some(MemoryEmbeddingApiConfig {
-        model: model.to_string(),
-        api_base,
-        api_key: api_key.to_string(),
-    })
-}
-
-async fn embed_text_vectors_via_api(
-    inputs: &[String],
-    dimensions: usize,
-    config: &MemoryEmbeddingApiConfig,
-) -> Result<Vec<Vec<f32>>, String> {
-    if inputs.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    let client = reqwest::Client::new();
-    let response = client
-        .post(format!("{}/embeddings", config.api_base))
-        .bearer_auth(config.api_key.as_str())
-        .json(&json!({
-            "model": config.model,
-            "input": inputs,
-        }))
-        .send()
-        .await
-        .map_err(|error| format!("embedding request failed: {error}"))?;
-
-    if !response.status().is_success() {
-        let status = response.status();
-        let body = response.text().await.unwrap_or_default();
-        return Err(format!(
-            "embedding request failed with status {}: {}",
-            status.as_u16(),
-            truncate_chars(&body, 240)
-        ));
-    }
-
-    let payload = response
-        .json::<Value>()
-        .await
-        .map_err(|error| format!("failed to parse embedding response json: {error}"))?;
-    let data = payload
-        .get("data")
-        .and_then(Value::as_array)
-        .ok_or_else(|| "embedding response missing data array".to_string())?;
-    if data.len() != inputs.len() {
-        return Err(format!(
-            "embedding response size mismatch: expected {}, got {}",
-            inputs.len(),
-            data.len()
-        ));
-    }
-
-    let mut vectors = Vec::with_capacity(data.len());
-    for item in data {
-        let raw_embedding = item
-            .get("embedding")
-            .and_then(Value::as_array)
-            .ok_or_else(|| "embedding item missing embedding array".to_string())?;
-        let parsed = raw_embedding
-            .iter()
-            .map(|component| {
-                component
-                    .as_f64()
-                    .map(|value| value as f32)
-                    .ok_or_else(|| "embedding component must be numeric".to_string())
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-        vectors.push(resize_and_normalize_embedding(&parsed, dimensions));
-    }
-
-    Ok(vectors)
-}
-
-fn resize_and_normalize_embedding(values: &[f32], dimensions: usize) -> Vec<f32> {
-    let dimensions = dimensions.max(1);
-    let mut resized = vec![0.0f32; dimensions];
-    for (index, value) in values.iter().enumerate() {
-        let bucket = index % dimensions;
-        resized[bucket] += *value;
-    }
-
-    let magnitude = resized
-        .iter()
-        .map(|component| component * component)
-        .sum::<f32>()
-        .sqrt();
-    if magnitude > 0.0 {
-        for component in &mut resized {
-            *component /= magnitude;
-        }
-    }
-    resized
-}
-
-fn build_structured_output_retry_prompt(schema: &Value, error: &str) -> String {
-    let schema_text = serde_json::to_string(schema).unwrap_or_else(|_| schema.to_string());
-    format!(
-        "Your previous response could not be accepted as structured JSON ({error}). \
-Please reply with only valid JSON that matches this schema exactly:\n{schema_text}"
-    )
-}
-
-fn parse_structured_output(messages: &[Message], schema: &Value) -> Result<Value, AgentError> {
-    let assistant = messages
-        .iter()
-        .rev()
-        .find(|message| message.role == MessageRole::Assistant)
-        .ok_or_else(|| {
-            AgentError::StructuredOutput(
-                "assistant response missing for structured output".to_string(),
-            )
-        })?;
-    let content = assistant.text_content();
-    let value = extract_json_payload(&content).map_err(AgentError::StructuredOutput)?;
-    validate_json_against_schema(schema, &value).map_err(AgentError::StructuredOutput)?;
-    Ok(value)
-}
-
-fn extract_json_payload(text: &str) -> Result<Value, String> {
-    let trimmed = text.trim();
-    if trimmed.is_empty() {
-        return Err("assistant response was empty; expected JSON output".to_string());
-    }
-
-    if let Ok(value) = serde_json::from_str::<Value>(trimmed) {
-        return Ok(value);
-    }
-
-    let mut cursor = 0usize;
-    while let Some(open_rel) = text[cursor..].find("```") {
-        let open = cursor + open_rel;
-        let after_open = &text[open + 3..];
-        let header_end_rel = after_open.find('\n').unwrap_or(after_open.len());
-        let header = after_open[..header_end_rel].trim();
-        let block_start = if header_end_rel < after_open.len() {
-            open + 3 + header_end_rel + 1
-        } else {
-            open + 3 + header_end_rel
-        };
-        let Some(close_rel) = text[block_start..].find("```") else {
-            break;
-        };
-        let close = block_start + close_rel;
-        cursor = close + 3;
-
-        if !(header.is_empty() || header.eq_ignore_ascii_case("json")) {
-            continue;
-        }
-
-        let block = text[block_start..close].trim();
-        if block.is_empty() {
-            continue;
-        }
-        if let Ok(value) = serde_json::from_str::<Value>(block) {
-            return Ok(value);
-        }
-    }
-
-    Err("assistant response did not contain parseable JSON content".to_string())
-}
-
-fn validate_json_against_schema(schema: &Value, payload: &Value) -> Result<(), String> {
-    let validator = validator_for(schema)
-        .map_err(|error| format!("invalid structured output schema: {error}"))?;
-    let mut errors = validator.iter_errors(payload);
-    if let Some(first) = errors.next() {
-        return Err(format!(
-            "structured output schema validation failed: {first}"
-        ));
-    }
-    Ok(())
-}
-
-async fn execute_tool_call_inner(
-    call: ToolCall,
-    registered: Option<(ToolDefinition, Arc<dyn AgentTool>)>,
-    tool_timeout: Option<Duration>,
-    cancellation_token: Option<CooperativeCancellationToken>,
-) -> ToolExecutionResult {
-    if cancellation_token
-        .as_ref()
-        .map(CooperativeCancellationToken::is_cancelled)
-        .unwrap_or(false)
-    {
-        return ToolExecutionResult::error(json!({
-            "error": format!("tool '{}' cancelled before execution", call.name)
-        }));
-    }
-
-    if let Some((definition, tool)) = registered {
-        if let Err(error) = validate_tool_arguments(&definition, &call.arguments) {
-            return ToolExecutionResult::error(json!({ "error": error }));
-        }
-
-        let tool_name = definition.name.clone();
-        let execution = async move {
-            if let Some(timeout) = tool_timeout {
-                match tokio::time::timeout(timeout, tool.execute(call.arguments)).await {
-                    Ok(result) => result,
-                    Err(_) => ToolExecutionResult::error(json!({
-                        "error": format!(
-                            "tool '{}' timed out after {}ms",
-                            tool_name,
-                            timeout.as_millis()
-                        )
-                    })),
-                }
-            } else {
-                tool.execute(call.arguments).await
-            }
-        };
-
-        if let Some(token) = cancellation_token {
-            tokio::select! {
-                _ = token.cancelled() => ToolExecutionResult::error(json!({
-                    "error": format!("tool '{}' cancelled", definition.name)
-                })),
-                result = execution => result,
-            }
-        } else {
-            execution.await
-        }
-    } else {
-        ToolExecutionResult::error(json!({
-            "error": format!("Tool '{}' is not registered", call.name)
-        }))
-    }
-}
-
-async fn sleep_with_cancellation(
-    delay: Duration,
-    cancellation_token: Option<CooperativeCancellationToken>,
-) -> Result<(), AgentError> {
-    if let Some(token) = cancellation_token {
-        tokio::select! {
-            _ = token.cancelled() => Err(AgentError::Cancelled),
-            _ = tokio::time::sleep(delay) => Ok(()),
-        }
-    } else {
-        tokio::time::sleep(delay).await;
-        Ok(())
-    }
-}
-
-fn spawn_async_event_handler_worker(
-    receiver: std::sync::mpsc::Receiver<AgentEvent>,
-    handler: AsyncEventHandler,
-    timeout: Option<Duration>,
-    metrics: Arc<AsyncEventDispatchMetricsInner>,
-) {
-    std::thread::spawn(move || {
-        let runtime = match tokio::runtime::Builder::new_current_thread()
-            .enable_time()
-            .build()
-        {
-            Ok(runtime) => runtime,
-            Err(_) => return,
-        };
-
-        while let Ok(event) = receiver.recv() {
-            let handler = Arc::clone(&handler);
-            let metrics = Arc::clone(&metrics);
-            runtime.block_on(async move {
-                let mut task = tokio::spawn(async move { (handler)(event).await });
-                if let Some(timeout) = timeout {
-                    match tokio::time::timeout(timeout, &mut task).await {
-                        Ok(Ok(())) => {
-                            metrics.completed.fetch_add(1, Ordering::Relaxed);
-                        }
-                        Ok(Err(_)) => {
-                            metrics.panicked.fetch_add(1, Ordering::Relaxed);
-                        }
-                        Err(_) => {
-                            task.abort();
-                            let _ = task.await;
-                            metrics.timed_out.fetch_add(1, Ordering::Relaxed);
-                        }
-                    }
-                } else {
-                    match task.await {
-                        Ok(()) => {
-                            metrics.completed.fetch_add(1, Ordering::Relaxed);
-                        }
-                        Err(_) => {
-                            metrics.panicked.fetch_add(1, Ordering::Relaxed);
-                        }
-                    }
-                }
-            });
-        }
-    });
-}
-
-fn lock_or_recover<T>(mutex: &Mutex<T>) -> std::sync::MutexGuard<'_, T> {
-    match mutex.lock() {
-        Ok(guard) => guard,
-        Err(poisoned) => poisoned.into_inner(),
-    }
-}
-
-fn cache_insert_with_limit<T: Clone>(
-    cache: &mut HashMap<String, T>,
-    order: &mut VecDeque<String>,
-    key: String,
-    value: T,
-    max_entries: usize,
-) {
-    if max_entries == 0 {
-        return;
-    }
-    if let Some(position) = order.iter().position(|entry| entry == &key) {
-        order.remove(position);
-    }
-    order.push_back(key.clone());
-    cache.insert(key, value);
-
-    while cache.len() > max_entries {
-        let Some(oldest) = order.pop_front() else {
-            break;
-        };
-        cache.remove(&oldest);
-    }
-}
-
-#[derive(Default)]
-struct StreamingRetryBufferState {
-    delivered_output: String,
-    attempt_output: String,
-}
-
-impl StreamingRetryBufferState {
-    fn reset_attempt(&mut self) {
-        self.attempt_output.clear();
-    }
-}
-
-fn stream_retry_buffer_on_delta(
-    state: &mut StreamingRetryBufferState,
-    delta: &str,
-) -> Option<String> {
-    state.attempt_output.push_str(delta);
-    if state.delivered_output.is_empty() {
-        state.delivered_output.push_str(delta);
-        return Some(delta.to_string());
-    }
-
-    if state.attempt_output.len() <= state.delivered_output.len() {
-        if state.delivered_output.starts_with(&state.attempt_output) {
-            return None;
-        }
-        state.delivered_output.push_str(delta);
-        return Some(delta.to_string());
-    }
-
-    if state.attempt_output.starts_with(&state.delivered_output) {
-        let replay = state
-            .attempt_output
-            .get(state.delivered_output.len()..)
-            .unwrap_or_default()
-            .to_string();
-        if replay.is_empty() {
-            return None;
-        }
-        state.delivered_output.push_str(&replay);
-        return Some(replay);
-    }
-
-    state.delivered_output.push_str(delta);
-    Some(delta.to_string())
-}
-
-fn timeout_duration_from_ms(timeout_ms: Option<u64>) -> Option<Duration> {
-    timeout_ms
-        .filter(|timeout_ms| *timeout_ms > 0)
-        .map(Duration::from_millis)
-}
-
-fn is_retryable_ai_error(error: &TauAiError) -> bool {
-    match error {
-        TauAiError::Http(http) => http.is_timeout() || http.is_connect(),
-        TauAiError::HttpStatus { status, .. } => {
-            *status == 408 || *status == 409 || *status == 425 || *status == 429 || *status >= 500
-        }
-        TauAiError::MissingApiKey | TauAiError::Serde(_) | TauAiError::InvalidResponse(_) => false,
-    }
-}
-
-fn validate_tool_arguments(definition: &ToolDefinition, arguments: &Value) -> Result<(), String> {
-    let validator = validator_for(&definition.parameters)
-        .map_err(|error| format!("invalid JSON schema for '{}': {error}", definition.name))?;
-
-    let mut errors = validator.iter_errors(arguments);
-    if let Some(first) = errors.next() {
-        return Err(format!(
-            "invalid arguments for '{}': {}",
-            definition.name, first
-        ));
-    }
-
-    Ok(())
-}
-
-fn normalize_direct_message_content(
-    content: &str,
-    max_message_chars: usize,
-) -> Result<String, AgentDirectMessageError> {
-    let normalized = content.trim();
-    if normalized.is_empty() {
-        return Err(AgentDirectMessageError::EmptyContent);
-    }
-    let actual_chars = normalized.chars().count();
-    if actual_chars > max_message_chars {
-        return Err(AgentDirectMessageError::MessageTooLong {
-            actual_chars,
-            max_chars: max_message_chars,
-        });
-    }
-    Ok(normalized.to_string())
 }
 
 #[cfg(test)]
